+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-dashboard-sl` will be documented in this file
 
+## 2.0.0 - 2020-09-09
+
+- Updated to spatie/laravel-dashboard 2.0
+- Added config to remove early departures
+
 ## 1.0.0 - 2020-05-04
 
 - initial release